% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_analyzer.R
\name{calculate_weekly_stats}
\alias{calculate_weekly_stats}
\title{Calculate weekly stats}
\usage{
calculate_weekly_stats(
  df,
  stat_col,
  filter_id = NULL,
  id_col = "Id",
  date_col = "ActivityDate"
)
}
\arguments{
\item{df}{Fitbit data as a DataFrame}

\item{stat_col}{DataFrame statistic column name to be aggregated}

\item{filter_id}{If specified, will generate stats for only this Id}

\item{id_col}{Column name that uniquely identifies a user; default = "Id"}

\item{date_col}{Column name for the statistic date/time; default = "ActivityDate"}
}
\value{
The weekly aggregated data with total_ and avg_ prefixed columns.
}
\description{
Aggregates stats to create weekly totals and averages. This method can be 
used to aggregate the data for multiple users, or can be filtered for a 
single user.
}
