% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_analyzer.R
\name{plot_weekly_trend}
\alias{plot_weekly_trend}
\title{Plot weekly trend}
\usage{
plot_weekly_trend(
  df,
  stat_col,
  filter_id,
  id_col = "Id",
  date_col = "ActivityDate"
)
}
\arguments{
\item{df}{Fitbit data as a DataFrame}

\item{stat_col}{DataFrame statistic column name to be tracked}

\item{filter_id}{User Id for which to generate the trendline}

\item{id_col}{Column name that uniquely identifies a user; default = "Id"}

\item{date_col}{Column name for the statistic date/time; default = "ActivityDate"}
}
\value{
A list of plots of the total and average weekly trend, respectively.
}
\description{
Calculates the weekly stats for the specified column and plots it against the 
week number to provide a picture of the trend.
}
