% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_analyzer.R
\name{read_fitbit_data}
\alias{read_fitbit_data}
\title{Read Fitbit data}
\usage{
read_fitbit_data(file)
}
\arguments{
\item{file}{Location of the exported CSV file}
}
\value{
The data as a DataFrame
}
\description{
Reads the CSV generated by exporting the data from Fitbit
}
